+++ main.go
@@ -6,28 +6,39 @@ import (
 	"strings"
 
 	"github.com/marpaia/graphite-golang"
+	"github.com/sensu-community/sensu-plugin-sdk/sensu"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
-	"github.com/sensu/sensu-plugins-go-library/sensu"
 )
 
 type HandlerConfig struct {
 	sensu.PluginConfig
-	Prefix   string
-	Port     uint64
-	Host     string
-	NoPrefix bool
+	Prefix           string
+	PrefixSource     bool
+	AnnotationPrefix string
+	Port             uint64
+	Host             string
+	NoPrefix         bool
 }
 
 const (
-	prefix      = "prefix"
-	port        = "port"
-	host        = "host"
-	noPrefix    = "no-prefix"
-	defaultPort = 2003
+	// flags
+	prefix           = "prefix"
+	prefixSource     = "prefix-source"
+	annotationPrefix = "annotation-prefix"
+	port             = "port"
+	host             = "host"
+	noPrefix         = "no-prefix"
+
+	// defaults
+	defaultPrefix           = "sensu"
+	defaultPort             = 2003
+	defaultHost             = "127.0.0.1"
+	defaultAnnotationPrefix = "sensu.annotations"
 )
 
 var (
-	stdin *os.File
+	stdin        *os.File
+	metricPrefix string
 
 	config = HandlerConfig{
 		PluginConfig: sensu.PluginConfig{
@@ -39,12 +50,28 @@ var (
 
 	graphiteConfigOptions = []*sensu.PluginConfigOption{
 		{
+			Path:      prefixSource,
+			Argument:  prefixSource,
+			Shorthand: "s",
+			Default:   false,
+			Usage:     "If true, prepends the sensu entity name (source) as a string to each metric in graphite",
+			Value:     &config.PrefixSource,
+		},
+		{
 			Path:      prefix,
 			Argument:  prefix,
 			Shorthand: "P",
-			Default:   "sensu",
-			Usage:     "The prefix to use in graphite for these metrics",
+			Default:   defaultPrefix,
+			Usage:     "The string to be prepended to each metric in graphite",
 			Value:     &config.Prefix,
+		},
+		{
+			Path:      annotationPrefix,
+			Argument:  annotationPrefix,
+			Shorthand: "a",
+			Default:   defaultAnnotationPrefix,
+			Usage:     "The string to be prepended to each annotation in graphite",
+			Value:     &config.AnnotationPrefix,
 		},
 		{
 			Path:      port,
@@ -58,7 +85,7 @@ var (
 			Path:      host,
 			Argument:  host,
 			Shorthand: "H",
-			Default:   "localhost",
+			Default:   defaultHost,
 			Usage:     "The hostname or address of the graphite server",
 			Value:     &config.Host,
 		},
@@ -67,7 +94,7 @@ var (
 			Argument:  noPrefix,
 			Shorthand: "n",
 			Default:   false,
-			Usage:     "Do not include *any* prefixes, use the bare metrics.point.name",
+			Usage:     "Unsets the default prefix value, use the bare metrics.point.name",
 			Value:     &config.NoPrefix,
 		},
 	}
@@ -78,10 +105,22 @@ func main() {
 	goHandler.Execute()
 }
 
-func checkArgs(_ *corev2.Event) error {
+func checkArgs(event *corev2.Event) error {
+	if !event.HasMetrics() {
+		return fmt.Errorf("event does not contain metrics")
+	}
 
-	if config.NoPrefix {
-		config.Prefix = ""
+	prefixSource := strings.Replace(event.Entity.Name, ".", "_", 1)
+	if config.Prefix != "" && !config.NoPrefix {
+		// --prefix is set, --no-prefix is not set
+		metricPrefix = config.Prefix
+		if config.PrefixSource {
+			// --prefix and --prefix-source are both set
+			metricPrefix = fmt.Sprintf("%s.%s", metricPrefix, prefixSource)
+		}
+	} else if config.PrefixSource {
+		// --prefix-source and --no-prefix are both set
+		metricPrefix = prefixSource
 	}
 
 	if stdin == nil {
@@ -92,38 +131,50 @@ func checkArgs(_ *corev2.Event) error {
 }
 
 func sendMetrics(event *corev2.Event) error {
-
 	var (
-		metrics        []graphite.Metric
-		tmp_point_name string
+		metrics      []graphite.Metric
+		tmpPointName string
+		tmpname      string
+		check        string
 	)
 
-	Graphite, err := graphite.NewGraphiteWithMetricPrefix(config.Host, int(config.Port), config.Prefix)
+	entity := strings.Replace(event.Entity.Name, ".", "_", 1)
+	g, err := graphite.NewGraphite(config.Host, int(config.Port))
 	if err != nil {
 		return err
 	}
 
 	for _, point := range event.Metrics.Points {
-		if config.NoPrefix {
-			tmpvalue := fmt.Sprintf("%f", point.Value)
-			metrics = append(metrics, graphite.NewMetric(point.Name, tmpvalue, point.Timestamp))
+		if metricPrefix == "" {
+			metrics = append(metrics, graphite.NewMetric(point.Name, fmt.Sprintf("%f", point.Value), point.Timestamp))
 		} else {
 			// Deal with special cases, such as disk checks that return file system paths as the name
 			// Graphite places these on disk using the name, so using any slashes would cause confusion and lost metrics
 			if point.Name == "/" {
-				tmp_point_name = "root"
+				tmpPointName = "root"
 			} else {
-				tmp_point_name = strings.Replace(point.Name, "/", "_", -1)
+				tmpPointName = strings.Replace(point.Name, "/", "_", -1)
 			}
-			tmpname := fmt.Sprintf("%s.%s_%s", strings.Replace(event.Entity.Name, ".", "_", 1), event.Check.Name, tmp_point_name)
-			tmpvalue := fmt.Sprintf("%f", point.Value)
-			metrics = append(metrics, graphite.NewMetric(tmpname, tmpvalue, point.Timestamp))
+
+			if event.HasCheck() {
+				check = event.Check.Name
+			} else {
+				check = "statsd"
+			}
+			tmpname = fmt.Sprintf("%s.%s.%s_%s", metricPrefix, entity, check, tmpPointName)
+			metrics = append(metrics, graphite.NewMetric(tmpname, fmt.Sprintf("%f", point.Value), point.Timestamp))
 		}
 	}
 
-	if err = Graphite.SendMetrics(metrics); err != nil {
+	if event.HasCheck() {
+		annotationPath := fmt.Sprintf("%s.events.%s.%s", config.AnnotationPrefix, entity, event.Check.Name)
+		metrics = append(metrics, graphite.NewMetric(fmt.Sprintf("%s.action.%s", annotationPath, event.Check.State), "1", event.Timestamp))
+		metrics = append(metrics, graphite.NewMetric(fmt.Sprintf("%s.status", annotationPath), fmt.Sprintf("%d", event.Check.Status), event.Timestamp))
+	}
+
+	if err = g.SendMetrics(metrics); err != nil {
 		return err
 	}
 
-	return Graphite.Disconnect()
+	return g.Disconnect()
 }